pyproject.toml
[project]
name = "invoice-parser"
version = "0.1.0"
dependencies = [
    "instructor>=1.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
py-modules = ["client", "invoices", "main", "models", "parser"]
